feat : 스터디룸 관련 crud 추가

+++ backend/src/main/java/com/ssafy/ptpt/api/studyroom/request/StudyRoomConnectRequest.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class StudyRoomConnectRequest {
-    private Integer studyRoomId;
+    private Long studyRoomId;
     private String studyRoomPw;
 }
